flow: split static flow in its own header file
