+++ src/test/test-flow.c
@@ -31,6 +31,7 @@
  */
 
 #include "sol-flow.h"
+#include "sol-flow-static.h"
 #include "sol-mainloop.h"
 #include "sol-util.h"
 #include "sol-vector.h"
